Tweaking build system for modules depending on HDF5 for Windows
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@6868 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/interpolate1d/interpolate1d/interpolate1d.cpp
@@ -93,7 +93,8 @@ namespace interpolate1d
     hpx::naming::id_type interpolate1d::get_gid(double value)
     {
         std::size_t partition_size = num_elements_ / partitions_.size();
-        std::size_t index = (value - minval_) / (delta_ * partition_size);
+        std::size_t index = static_cast<std::size_t>(
+            (value - minval_) / (delta_ * partition_size));
 
         if (index == partitions_.size())
             --index;
